replace old PostCard with PostCardNeo
